% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results}
\alias{results}
\title{results}
\format{
A tibble with 4027 rows and 12 variables:
\describe{
\item{round_number}{Indicates the round number (numeric)}
\item{round}{Indicates the round (character)}
\item{board}{The first number indicates the position of the table.
Table 1 is where the leading team plays its matches.
The second number is the player's position in the team, for example.
1.3 indicates that the match happened in table 1,
and the player was the number three. }
\item{title_white}{Indicates the FIDE title of the player with white pieces.
For example, GM= Grand Master, IM= International Master.}
\item{name_white}{The name of the player with white pieces}
\item{elo_white}{The ELO (a measure of the strength of the player)
of the player with white pieces}
\item{title_black}{Indicates the FIDE title of the player with black pieces.
For example, GM= Grand Master, IM= International Master.}
\item{name_black}{The name of the player with black pieces}
\item{elo_black}{The ELO (a measure of the strength of the player)
of the player with black pieces}
\item{result_white}{The final result of the match from the perspective
of the player with the white pieces (won, lost, or draw)}
\item{result_black}{The final result of the match from the perspective
of the player with the black pieces (won, lost, or draw)}
\item{elo_difference}{The result of subtracting
the ELO of the two players. ELO player 1 - ELO player 2.
A positive difference indicates that player 1 was stronger}
}
}
\usage{
data(results)
}
\description{
describe your data set here
}
\keyword{datasets}
