% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\docType{data}
\name{teams}
\alias{teams}
\title{teams}
\format{
A tibble with 6 rows and 6 variables:
\describe{\item{no}{label here}
\item{fed}{label here}
\item{team}{label here}
\item{country}{label here}
\item{rtg_avg}{label here}
\item{captain}{label here}
}
}
\usage{
data(teams)
}
\description{
describe your data set here
}
\keyword{datasets}
