% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\docType{data}
\name{players}
\alias{players}
\title{players}
\format{
A tibble with 11 rows and 11 variables:
\describe{
\item{No_player}{label here}
\item{title}{label here}
\item{name}{label here}
\item{fide_id}{label here}
\item{rtg}{label here}
\item{fed}{label here}
\item{team}{label here}
\item{Board}{label here}
\item{performance_rating}{label here}
\item{k}{label here}
\item{change_rating}{label here}
}
}
\usage{
data(players)
}
\description{
describe your data set here
}
\keyword{datasets}
