% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results}
\alias{results}
\title{results}
\format{
A tibble with 12 rows and 12 variables:
\describe{\item{round_number}{label here}
\item{round}{label here}
\item{board}{label here}
\item{title_white}{label here}
\item{name_white}{label here}
\item{elo_white}{label here}
\item{title_black}{label here}
\item{name_black}{label here}
\item{elo_black}{label here}
\item{result_white}{label here}
\item{result_black}{label here}
\item{elo_difference}{label here}
}
}
\usage{
data(results)
}
\description{
describe your data set here
}
\keyword{datasets}
