% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\docType{data}
\name{teams}
\alias{teams}
\title{teams}
\format{
A tibble with 188 rows and 6 variables:
\describe{
\item{no}{the number of the team.
Team 1 is the strongest (highest average ELO) team in the tournament}
\item{fed}{Three letters representing the federation}
\item{team}{Three letters representing the country's team}
\item{country}{A character with the name of the country}
\item{rtg_avg}{The rating average.
This number is the average ELO of the team}
\item{captain}{The name of the captain of the team}
}
}
\usage{
data(teams)
}
\description{
describe your data set here
}
\keyword{datasets}
